refactor(ai-diagnostic): split son_functions into models + optimizers
son_functions.py had no coverage. Added 7 characterisation tests (MLB offload
scenarios, recommendation serialisation, enum vocabulary, SONEngine end-to-end
and function-filtered runs), then extracted, test-first:

  service/son_models.py     - function/status/priority enums + CellMetrics,
                              SONRecommendation
  service/son_optimizers.py - MLB / MRO / CCO / EnergySaving optimizers

son_functions.py keeps the SONEngine orchestration + module API and re-exports
every name (importers unchanged). Dropped now-dead imports (numpy, random,
deque, dataclass/field, Enum, Set/Tuple, timedelta). 752 -> 273 LOC. All 208
ai-diagnostic tests pass (201 + 7 new).
